refactor: extract template form schema into a separate file and update form handling in template form component
